% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{play_audio}
\alias{play_audio}
\title{Play audio in a browser}
\usage{
play_audio(audio = "output.wav", html = "player.html")
}
\arguments{
\item{audio}{The file location of the audio file.  Must be supported by HTML5}

\item{html}{The html file location that will be created host the audio}
}
\description{
This uses HTML5 audio tags to play audio in your browser.
}
\details{
Borrowed from \code{googleLanguageR::gl_talk_player()}
}
