% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2speech-package.R
\docType{package}
\name{text2speech-package}
\alias{text2speech}
\alias{text2speech-package}
\title{text2speech: Text to Speech Conversion}
\description{
Converts text into speech using various text-to-speech (TTS) engines and provides an unified interface for accessing their functionality. With this package, users can easily generate audio files of spoken words, phrases, or sentences from plain text data. The package supports multiple TTS engines, including Google's 'Cloud Text-to-Speech API', 'Amazon Polly', Microsoft's 'Cognitive Services Text to Speech REST API', the 'Speechify Text-to-Speech API', and a free TTS engine called 'Coqui TTS'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jhudsl/text2speech}
  \item Report bugs at \url{https://github.com/jhudsl/text2speech/issues}
}

}
\author{
\strong{Maintainer}: John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID})

Authors:
\itemize{
  \item John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID})
  \item Howard Baek \email{howardbaek.fh@gmail.com} (\href{https://orcid.org/0009-0000-8942-1618}{ORCID})
}

}
\keyword{internal}
