% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_backend.R
\name{tts_coqui}
\alias{tts_coqui}
\title{Convert Text to Speech using Coqui TTS}
\usage{
tts_coqui(
  text,
  exec_path,
  output_format = c("wav", "mp3"),
  model_name = "tacotron2-DDC_ph",
  vocoder_name = "ljspeech/univnet",
  bind_audio = TRUE,
  save_local = FALSE,
  save_local_dest = NULL,
  ...
)
}
\arguments{
\item{text}{A character vector of text to be spoken}

\item{exec_path}{System path to Coqui TTS}

\item{output_format}{Format of output files: "mp3" or "wav"}

\item{model_name}{Deep Learning model for Text-to-Speech Conversion}

\item{vocoder_name}{Model that generates audio}

\item{bind_audio}{Should the \code{\link[=tts_bind_wav]{tts_bind_wav()}} be run on after
the audio has been created, to ensure that the length of text and the
number of rows is consistent? This affects the output format of some audio.
description}

\item{save_local}{Should the output file be saved locally?}

\item{save_local_dest}{Destination to save output file, if saved locally}

\item{...}{Additional arguments}
}
\description{
Convert Text to Speech using Coqui TTS
}
