% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_coqui_installed.R
\name{tts_coqui_installed}
\alias{tts_coqui_installed}
\title{Is coqui TTS installed on local system?}
\usage{
tts_coqui_installed()
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Is coqui TTS installed on local system?
}
