% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_voices.R
\name{tts_coqui_voices}
\alias{tts_coqui_voices}
\title{Get Coqui TTS voices}
\usage{
tts_coqui_voices()
}
\value{
A \code{data.frame} of the language, dataset, and model name.
}
\description{
Get Coqui TTS voices
}
