% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_voices.R
\name{tts_default_voice}
\alias{tts_default_voice}
\title{Default voice for text-to-speech engine}
\usage{
tts_default_voice(
  service = c("amazon", "google", "microsoft", "coqui", "speechify")
)
}
\arguments{
\item{service}{Text-to-speech engine}
}
\description{
Default voice for text-to-speech engine
}
