% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_speak_engine.R
\name{tts_speak_engine}
\alias{tts_speak_engine}
\title{Knitr Speak Enginge}
\usage{
tts_speak_engine(options)
}
\arguments{
\item{options}{A list of chunk options. Usually this is just the
object options passed to the engine function; see
\code{\link{knit_engines}}}
}
\value{
A character string generated from the source code and
output using the appropriate output hooks.
}
\description{
Knitr Speak Enginge
}
\examples{
knitr::knit_engines$set(speak = tts_speak_engine)
}
