% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_to_wav.R
\name{pcm_to_wav}
\alias{pcm_to_wav}
\title{Convert PCM to Wav}
\usage{
pcm_to_wav(input, output = tempfile(fileext = ".wav"),
  extensible = FALSE)
}
\arguments{
\item{input}{output from \link[aws.polly:get_synthesis]{aws.polly::get_synthesis} or PCM filename}

\item{output}{output file for Wav file}

\item{extensible}{passed to \link[tuneR:writeWave]{tuneR::writeWave}}
}
\value{
A filename of the output
}
\description{
Convert PCM to Wav
}
