% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{set_coqui_path}
\alias{set_coqui_path}
\title{Point to local coqui tts Executable File}
\usage{
set_coqui_path(path)
}
\arguments{
\item{path}{path to the local \code{coqui} tts Executable File}
}
\value{
Returns nothing, function sets the option variable
\code{path_to_coqui}.
}
\description{
Function to set an option that points to the local \code{coqui} tts Executable File
\code{tts}.
}
\details{
List of possible file path locations for the local \code{coqui} tts
Executable File
\describe{
\item{Linux}{\verb{/usr/bin/tts}, \verb{/usr/local/bin/tts}}
\item{Mac}{\verb{/opt/homebrew/Caskroom/miniforge/base/bin/tts}}
\item{Windows}{\verb{C:\\\\Program Files\\\\tts}}
}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
set_coqui_path("~/path/to/tts")
\dontshow{\}) # examplesIf}
}
