% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_backend.R, R/tts_synthesize.R
\name{tts_google}
\alias{tts_google}
\alias{tts_amazon}
\alias{tts_microsoft}
\alias{tts}
\title{Text to Speech}
\usage{
tts_google(text, output_format = c("mp3", "wav"),
  voice = "en-US-Standard-C", ...)

tts_amazon(text, output_format = c("mp3", "wav"), voice = "Joanna",
  ...)

tts_microsoft(text, output_format = c("mp3", "wav"),
  voice = "Microsoft Server Speech Text to Speech Voice (en-US, ZiraRUS)",
  ...)

tts(text, output_format = c("mp3", "wav"), ..., service = c("amazon",
  "google", "microsoft"))
}
\arguments{
\item{text}{A character vector of text to speak}

\item{output_format}{Format of output files}

\item{voice}{A full voice name that can be passed to the
service, such as the
argument `voice` for [aws.polly::get_synthesis], or
or [mscstts::ms_synthesize()] or the
`name` argument for [googleLanguageR::gl_talk()]}

\item{...}{Additional arguments to
[text2speech::tts_google()],
[text2speech::tts_amazon()], or
[text2speech::tts_microsoft()]}

\item{service}{service to use}

\item{language}{An ISO 3166 country identification tag. If
NULL, then a data.frame of the results are returned.}
}
\value{
A `data.frame` of text and wav files
}
\description{
Text to Speech
}
\note{
All functions have a  `voice`` argument fro a
full voice name that can be passed to the
service, such as `voice` for [aws.polly::get_synthesis], or
}
