% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_amazon_auth.R, R/tts_auth.R,
%   R/tts_google_auth.R, R/tts_microsoft_auth.R
\name{tts_amazon_auth}
\alias{tts_amazon_auth}
\alias{tts_auth}
\alias{tts_google_authenticated}
\alias{tts_amazon_authenticated}
\alias{tts_microsoft_authenticated}
\alias{tts_google_auth}
\alias{tts_microsoft_auth}
\title{Authenticate the user's credentials for using the Amazon Polly Text-to-Speech (TTS) service}
\usage{
tts_amazon_auth(key_or_json_file = NULL, ...)

tts_auth(
  service = c("amazon", "google", "microsoft", "coqui"),
  key_or_json_file = NULL,
  ...
)

tts_google_authenticated()

tts_amazon_authenticated()

tts_microsoft_authenticated(...)

tts_google_auth(key_or_json_file = NULL, ...)

tts_microsoft_auth(key_or_json_file = NULL, ...)
}
\arguments{
\item{key_or_json_file}{Either an API key (for Microsoft)
or JSON file (for Google)}

\item{...}{Additional arguments to pass to
\code{\link{use_credentials}} or \code{\link{ms_get_tts_token}}}

\item{service}{type of synthesis engine}
}
\value{
A logical indicator of authorization
}
\description{
Authenticate the user's credentials for using the Amazon Polly Text-to-Speech (TTS) service

Authorize Text-to-Speech Engine

Check Google Cloud Text-to-Speech API Authentication Status

Check Amazon Polly Text-to-Speech API Authentication Status

Check Microsoft Cognitive Services Text to Speech REST API Authentication Status

Authenticate the user's Google Cloud credentials for using the Google Cloud Text-to-Speech (TTS) API

Authenticate the user's credentials for using the Microsoft Cognitive Services Text to Speech REST API,
}
\examples{
tts_auth("google")
tts_auth("microsoft")

tts_google_authenticated()
tts_microsoft_authenticated()

tts_google_auth()
tts_microsoft_auth()
if (requireNamespace("aws.polly", quietly = TRUE)) {
tts_auth("amazon")
tts_amazon_authenticated()
tts_amazon_auth()
}
}
