% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_voices.R
\name{tts_voices}
\alias{tts_voices}
\alias{tts_amazon_voices}
\alias{tts_microsoft_voices}
\alias{tts_google_voices}
\alias{tts_coqui_voices}
\title{Text to Speech Voices}
\usage{
tts_voices(service = c("amazon", "google", "microsoft", "coqui"), ...)

tts_amazon_voices(...)

tts_microsoft_voices(...)

tts_google_voices(...)

tts_coqui_voices()
}
\arguments{
\item{service}{service to use}

\item{...}{Additional arguments to service voice listings.}
}
\value{
A \code{data.frame} of language codes, voices, genders,
and language names
}
\description{
Text to Speech Voices

Get Amazon Polly TTS voices

Get Microsoft Cognitive Services Text to Speech voices

Get Google Cloud TTS voices

Get Coqui TTS voices (list models)
}
\examples{
if (tts_microsoft_auth()) {
tts_voices(service = "microsoft")
}
if (tts_google_auth()) {
tts_voices(service = "google")
}
if (requireNamespace("aws.polly", quietly = TRUE)) {
if (tts_amazon_auth()) {
tts_voices(service = "amazon")
}
}
}
